version 5 Command line 1: Start the script -> Result: Job started: ID of the job With the option that it sends an e-mail to "Maarten@youtic.com" when finished. This can be done with the notifye-mail function.
Command line 2:
Status of job ID -> The process of the job.

Command line 3:
Retrieve data -> Download all data available at the moment (So when it got "stuck" I can still download the data of the 5000 complete pages).
And than the script can continue to the formatting stage with the data.

+ Resume Job